subtracting 1 from the $port_id value in the egress_ids loop
This is to fix an off by one error.
The ERS reports VLAN membership using a bitmask, but the MSB is always 0.
So using the `q_bridge_bits2indices` function results in the reported ports membership being +1 wrong.
The best fix for this would be to create a new function exactly the same as `q_bridge_bits2indices` which subtracts the 1 before returning the array. 
I tried this, but it broke my install :)
Not sure how to add to the `functions.inc.php` file without breaking it.