Add more OIDs for determining hardware and serial attributes correctly (#14904)
Some FTD systems like 4110 and 4120 (probably more) use different OIDs (not .1 or .4) for that, so these were added. The version extension is probably not helpful, as the result is an empty string … but at least that OID exists.
Not: without this change, hardware+serial fields will be empty after a (re)discovery.